feat: Refactoring api clients package
