+++ src/main/java/com/iexapis/cloud/integration/concurrencystock/api/quotes/QuotesController.java
@@ -1,4 +1,4 @@
-package com.iexapis.cloud.integration.concurrencystock.in.rest;
+package com.iexapis.cloud.integration.concurrencystock.api.quotes;
 
 import com.iexapis.cloud.integration.concurrencystock.model.QuotesModel;
 import org.springframework.scheduling.annotation.Scheduled;
